# Conflicts:
#	src/main/java/com/surepark/cmu/facades/ParkingLotFacade.java
#	src/main/java/com/surepark/cmu/interfaces/ParkingLotInterface.java
#	src/main/resources/META-INF/mybatis/mapper/ParkingLotMapper.xml
#	target/m2e-wtp/web-resources/META-INF/maven/com.surepark.cmu/surepark-restful/pom.properties

+++ src/main/java/com/surepark/cmu/payment/PaymentAPI.java
@@ -9,7 +9,7 @@ public class PaymentAPI {
 	
 	
 	
-	public boolean payment(CardModel card, int money)
+	public boolean payment(CardModel card, double money)
 	{
 		boolean result = false;
